refactor: go forward and back block
